Rename Our Allies component data file from helpers.ts to data.ts
